+++ src/config/presets/app/health.rs
@@ -5,6 +5,7 @@
 /// Returns the default value for health path for preset.
 pub(super) fn default_health_path_for_preset(preset: &str) -> Option<&'static str> {
     match preset {
+        "frankenphp" | "reverb" | "soketi" => Some("/"),
         "laravel" => Some("/up"),
         "dusk" | "selenium" => Some("/wd/hub/status"),
         "gotenberg" => Some("/health"),
@@ -16,7 +17,8 @@ pub(super) fn default_health_path_for_preset(preset: &str) -> Option<&'static st
 /// Returns the default value for health statuses for preset.
 pub(super) fn default_health_statuses_for_preset(preset: &str) -> Option<Vec<u16>> {
     match preset {
-        "laravel" | "dusk" | "selenium" | "gotenberg" | "mailhog" | "mailpit" => Some(vec![200]),
+        "frankenphp" | "reverb" | "soketi" | "laravel" | "dusk" | "selenium" | "gotenberg"
+        | "mailhog" | "mailpit" => Some(vec![200]),
         _ => None,
     }
 }